config: fix GetLargestObjectID for pseudo object IDs
This was not behaving correctly, and was breaking things when a
pseudo table had the largest ID below MaxReservedDescID.

+++ pkg/keys/sql.go
@@ -217,7 +217,7 @@ func (d sqlDecoder) DecodeIndexPrefix(key roachpb.Key) ([]byte, uint32, uint32, 
 }
 
 // DecodeDescMetadataID decodes a descriptor ID from a descriptor metadata key.
-func (d sqlDecoder) DecodeDescMetadataID(key roachpb.Key) (uint64, error) {
+func (d sqlDecoder) DecodeDescMetadataID(key roachpb.Key) (uint32, error) {
 	// Extract table and index ID from key.
 	remaining, tableID, _, err := d.DecodeIndexPrefix(key)
 	if err != nil {
@@ -231,5 +231,5 @@ func (d sqlDecoder) DecodeDescMetadataID(key roachpb.Key) (uint64, error) {
 	if err != nil {
 		return 0, err
 	}
-	return id, nil
+	return uint32(id), nil
 }
